object api search deserves its own spec file
